refactor: 전체 코드 리팩토링 - 불필요한 주석 제거 - 사용되지 않는 변수 및 메서드 삭제 - for 문을 stream 으로 변경
