+++ src/main/java/subway/auth/dto/TokenRequest.java
@@ -1,6 +1,7 @@
 package subway.auth.dto;
 
 public class TokenRequest {
+
     private String email;
     private String password;
