+++ src/main/java/subway/auth/dto/TokenResponse.java
@@ -1,6 +1,7 @@
 package subway.auth.dto;
 
 public class TokenResponse {
+
     private String accessToken;
 
     public TokenResponse() {
