+++ src/main/java/subway/auth/infrastructure/JwtTokenProvider.java
@@ -8,6 +8,7 @@ import java.util.Date;
 
 @Component
 public class JwtTokenProvider {
+
     @Value("${security.jwt.token.secret-key}")
     private String secretKey;
     @Value("${security.jwt.token.expire-length}")
