+++ src/main/java/subway/line/domain/Line.java
@@ -5,14 +5,12 @@ import subway.station.domain.Station;
 import java.util.List;
 
 public class Line {
+
     private Long id;
     private String name;
     private String color;
     private int extraFare;
-    private Sections sections = new Sections();
-
-    public Line() {
-    }
+    private Sections sections;
 
     public Line(String name, String color, int extraFare) {
         this(0L, name, color, extraFare, new Sections());
@@ -24,10 +22,6 @@ public class Line {
 
     public Line(Long id, String name, String color) {
         this(id, name, color, 0, new Sections());
-    }
-
-    public Line(Long id, String name, String color, Sections sections) {
-        this(id, name, color, 0, sections);
     }
 
     public Line(Long id, String name, String color, int extraFare) {
@@ -64,11 +58,6 @@ public class Line {
 
     public List<Station> getStations() {
         return sections.getStations();
-    }
-
-    public void update(Line line) {
-        this.name = line.getName();
-        this.color = line.getColor();
     }
 
     public void addSection(Station upStation, Station downStation, int distance) {
