+++ src/main/java/subway/line/dto/LineRequest.java
@@ -1,6 +1,7 @@
 package subway.line.dto;
 
 public class LineRequest {
+
     private String name;
     private String color;
     private Long upStationId;
