+++ src/main/java/subway/member/application/MemberService.java
@@ -16,7 +16,6 @@ public class MemberService {
         this.memberDao = memberDao;
     }
 
-    @Transactional
     public MemberResponse createMember(MemberRequest request) {
         Member member = memberDao.insert(request.toMember());
         return MemberResponse.of(member);
@@ -40,12 +39,10 @@ public class MemberService {
         return MemberResponse.of(member);
     }
 
-    @Transactional
     public void updateMember(Long id, MemberRequest memberRequest) {
         memberDao.update(new Member(id, memberRequest.getEmail(), memberRequest.getPassword(), memberRequest.getAge()));
     }
 
-    @Transactional
     public void deleteMember(Long id) {
         memberDao.deleteById(id);
     }
