+++ src/main/java/subway/path/ui/PathController.java
@@ -14,17 +14,12 @@ import subway.station.application.StationService;
 
 @RestController
 public class PathController {
-    // TODO: 경로조회 기능 구현하기
 
     private PathService pathService;
-    private LineService lineService;
-    private StationService stationService;
 
     @Autowired
-    public PathController(PathService pathService, LineService lineService, StationService stationService) {
+    public PathController(PathService pathService) {
         this.pathService = pathService;
-        this.lineService = lineService;
-        this.stationService = stationService;
     }
 
     @GetMapping("/paths")
