+++ src/main/java/subway/station/domain/Station.java
@@ -3,6 +3,7 @@ package subway.station.domain;
 import java.util.Objects;
 
 public class Station {
+
     private Long id;
     private String name;
