+++ src/main/java/subway/station/dto/StationResponse.java
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class StationResponse {
+
     private Long id;
     private String name;
