+++ src/main/java/subway/station/ui/StationController.java
@@ -1,5 +1,6 @@
 package subway.station.ui;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -13,8 +14,10 @@ import java.util.List;
 
 @RestController
 public class StationController {
+
     private StationService stationService;
 
+    @Autowired
     public StationController(StationService stationService) {
         this.stationService = stationService;
     }
